fix: CEMAHandler compute incorrectly
