+++ pysatl_tsp/c/mahandler.c
@@ -1,6 +1,7 @@
 #include "handler.h"
 #include <stdio.h>
 #include <stdlib.h>
+#include "mahandler.h"
 
 /*
  * Circular queue insertion operation
