+++ pysatl_tsp/c/mahandler.h
@@ -1,0 +1,1 @@
+double tsp_op_MA(struct tsp_handler *handler, void *next);
